requirements.txt
aniso8601==8.0.0
autopep8==1.5
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.18.1
pandas==0.25.3
pycodestyle==2.5.0
PyMySQL==0.9.3
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
selenium==3.141.0
six==1.13.0
urllib3==1.25.7
Werkzeug==0.16.0
